cleanup auth and logging code
